feat(wallet): add support for importing BIP39 wallets
This commit adds support for importing BIP39 wallets. A new method `ImportBIP39Wallet` has been added to the `IWalletRepository` interface and implemented in the `WalletRepository` class. The `Wallet` and `Key` models have been updated to include a new boolean flag `IsBIP39Imported` to indicate that the wallet was imported from a BIP39 mnemonic. The `Key` model also includes a new boolean flag `IsBIP39ImportedKey` to indicate that the key was imported from a BIP39 mnemonic. The `NBXplorerService` has been updated to include two new methods `ScanUTXOSetAsync` and `GetScanUTXOSetInformationAsync` to support scanning the UTXO set for imported BIP39 wallets.