Removes restriction on number of keys count for wallet creation (#234)
* Removes restriction on number of keys count for wallet creation
* feat(LightningService.cs): add check for correct number of human signatures in multisig wallet
test(LightningServiceTests.cs): add test for incorrect number of human signatures in multisig wallet

The changes were made to ensure that the number of human signatures matches the number of signatures required for a multisig wallet. This is important for security reasons, as it prevents unauthorized transactions from being processed. The test was added to verify that the system correctly handles cases where the number of human signatures is incorrect.

---------

Co-authored-by: José A.P <53834183+Jossec101@users.noreply.github.com>